Rename "IdData" class to "SavedIdData"
