+++ src/general.js
@@ -2,7 +2,7 @@ import preact from 'preact';
 import I18nWidget, { I18nButton } from './Components/I18nWidget';
 import CommentsWidget from './Components/CommentsWidget';
 import Cookie from 'js-cookie';
-import IdData from './Utility/IdData';
+import SavedIdData from './Utility/SavedIdData';
 import t from './Utility/i18n';
 import Privacy, { PRIVACY_ACTIONS } from './Utility/Privacy';
